Implement automatic precompilation of the theme
Following the path of Adwaita, this commit implements the automation to
automatically precompile the GTK3 theme versions to gresource binary
files, to improve performance of theme rendering, and avoid unnecessary
re-read and re-parsing of the theme on every app run.

Given that the theme is coded in sass to begin with directly altering it
is not a reasonable thing to do, and so we don't loose anything by
having the installed theme in binary form, specially when accessing the
sources is easy enough.